mv SearchSourceBuilderTest to SearchSourceBuilderTests
